+++ Front/Login/membership.html
@@ -235,7 +235,6 @@
         }
 
     </script>
->>>>>>> main
 </body>
 
 </html>
