+++ Server/server.js
@@ -1,8 +1,6 @@
 const express = require('express');
 const bodyParser = require('body-parser');
-const bcrypt = require('bcrypt');
 const cors = require('cors');
-const mysql = require('mysql2');
 
 // 라우터 가져오기
 const userRoutes = require('./routes/users');
@@ -12,6 +10,8 @@ const tesksRoutes = require('./routes/tasks');
 const mentionsRoutes = require('./routes/mentions');
 const commentRoutes = require('./routes/comment');
 const notifiRoutes = require('./routes/notification');
+
+const emailRoutes = require('./routes/userVerification');
 
 const app = express();
 
@@ -29,6 +29,7 @@ app.use('/api/tasks', tesksRoutes);
 app.use('/api/mentions', mentionsRoutes);
 app.use('/api/comment', commentRoutes);
 app.use('/api/notification', notifiRoutes);
+app.use('/api/userVerification', emailRoutes);
 
 // 서버 시작
 const PORT = process.env.PORT || 3000;
